Serve dropper JS only if dropper is rendered
